reorganizing repo to be a general template for project organization
