+++ spatio_flux/__init__.py
@@ -4,9 +4,7 @@ TODO -- make a "register_types" function that takes a core, registers all types 
 """
 
 from process_bigraph import ProcessTypes
-
-# make type system
-core = ProcessTypes()
+from spatio_flux.processes import register_processes
 
 
 def apply_non_negative(schema, current, update, core):
@@ -17,13 +15,25 @@ def apply_non_negative(schema, current, update, core):
 positive_float = {
     '_type': 'positive_float',
     '_inherit': 'float',
-    '_apply': apply_non_negative
-}
-core.register('positive_float', positive_float)
+    '_apply': apply_non_negative}
+
 
 bounds_type = {
     'lower': 'maybe[float]',
-    'upper': 'maybe[float]'
-}
-core.register_process('bounds', bounds_type)
+    'upper': 'maybe[float]'}
 
+
+particle_type = {
+    'id': 'string',
+    'position': 'tuple[float,float]',
+    'size': 'float',
+    'local': 'map[float]',
+    'exchange': 'map[float]',    # {mol_id: delta_value}
+}
+
+def register_types(core):
+    core.register('positive_float', positive_float)
+    core.register('bounds', bounds_type)
+    core.register('particle', particle_type)
+    
+    return register_processes(core)
